sample: add OpenAMP sample application
This commit adds a sample application using OpenAMP for remote procedure
calls on the LPCXpresso54114. It is adapted from the RPMsg-Lite sample
application added in PR #5960, and uses the IPM driver to provide
interprocessor interrupts.

Instructions for how to compile the application, including OpenAMP and
libmetal is provided in the README.md file in the application's
toplevel directory.

Signed-off-by: Kristian Klomsten Skordal <kristian.skordal@nordicsemi.no>